2.84 commiting
i was an idiot with the 2.83 rant.  i blame not being quite used to BlueJ yet.  Netbeans took inputs in the same area as it's output, with a special code to read what was inputted, and it didn't require quotation marks to understand it.  though, all of the inputs for the code i did in there were numbers.